+++ cmd/tuple/delete.go
@@ -59,12 +59,12 @@ var deleteCmd = &cobra.Command{
 				return fmt.Errorf("failed to parse input tuples due to %w", err)
 			}
 
-			maxTuplesPerWrite, err := cmd.Flags().GetInt("max-tuples-per-write")
+			maxTuplesPerWrite, err := cmd.Flags().GetInt32("max-tuples-per-write")
 			if err != nil {
 				return fmt.Errorf("failed to parse max tuples per write due to %w", err)
 			}
 
-			maxParallelRequests, err := cmd.Flags().GetInt("max-parallel-requests")
+			maxParallelRequests, err := cmd.Flags().GetInt32("max-parallel-requests")
 			if err != nil {
 				return fmt.Errorf("failed to parse parallel requests due to %w", err)
 			}
@@ -99,8 +99,8 @@ var deleteCmd = &cobra.Command{
 func init() {
 	deleteCmd.Flags().String("file", "", "Tuples file")
 	deleteCmd.Flags().String("model-id", "", "Model ID")
-	deleteCmd.Flags().Int("max-tuples-per-write", MaxTuplesPerWrite, "Max tuples per write chunk.")
-	deleteCmd.Flags().Int("max-parallel-requests", MaxParallelRequests, "Max number of requests to issue to the server in parallel.") //nolint:lll
+	deleteCmd.Flags().Int32("max-tuples-per-write", MaxTuplesPerWrite, "Max tuples per write chunk.")
+	deleteCmd.Flags().Int32("max-parallel-requests", MaxParallelRequests, "Max number of requests to issue to the server in parallel.") //nolint:lll
 }
 
 func ExactArgsOrFlag(n int, flag string) cobra.PositionalArgs {